Add more styling throughout page
