+++ src/components/IngredientItem.js
